move to processing profiles to a separate file
